refactor: move node red settings to file
